refactor(auth): formating and update project
