+++ src/app/globals.css
@@ -1,5 +1,5 @@
 @import '@fontsource-variable/inter';
-@import "@fontsource-variable/roboto";
+@import '@fontsource-variable/roboto';
 @import '@ictroot/ui-kit/style.css';
 
 :root {
