+++ src/features/auth/index.ts
@@ -1,1 +1,2 @@
 export * from './sign-in'
+export * from './ui'
