+++ src/features/auth/sign-in/ui/SignInForm.module.scss
@@ -18,7 +18,6 @@
   }
 }
 
-
 .fields {
   display: flex;
   flex-direction: column;
