+++ src/lib/hooks.ts
@@ -1,6 +1,5 @@
-// lib/hooks.ts
 import { useDispatch, useSelector, useStore, TypedUseSelectorHook } from 'react-redux'
-import type { AppDispatch, RootState, AppStore } from '@/services/store'
+import { AppDispatch, AppStore, RootState } from '@/app/store'
 
 export const useAppDispatch: () => AppDispatch = useDispatch
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
