--- src/pages/info.txt
@@ -1,1 +1,0 @@
-Пока неизвестно нужна ли это инструктора папок
