+++ src/shared/ui/SVGComponents/index.ts
@@ -1,4 +1,1 @@
-export { GitHub, Google, BellOutline,
-  RussiaFlag,
-  Select,
-  UkFlag } from '@ictroot/ui-kit'
+export { GitHub, Google, BellOutline, RussiaFlag, Select, UkFlag } from '@ictroot/ui-kit'
